# fix the concurrency model of ResponseRouter that results in NullPoint…