@willsalz 2015-09-14
Damn. I should have left in that ConcurrentHash map…